Develop

+++ backend/routes/menu.py
@@ -5,9 +5,7 @@ from flask_jwt_extended import get_jwt_claims, jwt_required
 from app import api, db
 from db.menu_db import menu_DB
 
-import model.request_model as request_model
-import model.response_model as response_model
-from model.response_model import menu_items_model, menu_item_model
+import model.menu_model as menu_model
 
 menu_db = menu_DB(db)
 menu = api.namespace('menu', description='Menu Route')
@@ -15,15 +13,16 @@ menu = api.namespace('menu', description='Menu Route')
 @menu.route('/')
 class Menu(Resource):
     @jwt_required
-    @menu.response(200, 'Success')
+    @menu.response(200, 'Success', model = menu_model.menu_response_model)
     @menu.response(400, 'Invalid request')
     def get(self):
-        # Return the entire menu
+        # Get all the items in the menu
 
+        # Get all the categories in the menu
         categories = menu_db.get_categories()
-
         menu = []
 
+        # Query all the menu items in each category and append to the menu
         for category in categories:
             menu.append(menu_db.get_category(category['id']))
     
@@ -33,21 +32,43 @@ class Menu(Resource):
 @menu.route('/item')
 class CreateMenuItem(Resource):
     @jwt_required
-    @menu.response(200, 'Success')
+    @menu.response(200, 'Success', model=menu_model.menu_item_id_response_model)
     @menu.response(400, 'Invalid Request')
-    @menu.expect(request_model.menu_item_model)
+    @menu.response(500, 'Something went wrong')
+    @menu.expect(menu_model.menu_item_request_model)
     def post(self):
         # Create a new menu item
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
+        # Get and validate request body
         item = request.get_json()
+
+        if not(item.get('name') and item.get('price')):
+            abort(400, 'Request missing required fields')
+
+        if item.get('visible') is None:
+            item['visible'] = True
+        
+        if item.get('description') is None:
+            item['description'] = ''
+
+        if  item.get('image_url') is None:
+            item['image_url'] = ''
+
+
+        # Create new item in db
         item_id = menu_db.create_item(item)
         if (not item_id):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
         
         return jsonify({ 'item_id': item_id })
 
     @jwt_required
-    @menu.response(200, 'Success', model=response_model.menu_items_model)
-    @menu.response(400, 'Invalid Request')
+    @menu.response(200, 'Success', model=menu_model.menu_items_response_model)
     def get(self):
         # Get all menu items
         items = menu_db.get_all_menu_items()
@@ -58,9 +79,16 @@ class MenuItem(Resource):
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
-    @menu.expect(request_model.menu_item_model)
+    @menu.expect(menu_model.menu_item_request_model)
     def put(self, item_id):
         # Modify a menu item
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
+        # Get and validate request body
         edit = request.get_json()
         if (not menu_db.edit_item(edit, item_id)):
             abort(400, 'Something went wrong')
@@ -69,23 +97,30 @@ class MenuItem(Resource):
 
 
     @jwt_required
-    @menu.response(200, 'Success', model=response_model.menu_item_model)
-    @menu.response(400, 'Invalid Request')
+    @menu.response(200, 'Success', model=menu_model.menu_item_response_model)
+    @menu.response(500, 'Something went wrong')
     def get(self, item_id):
         # Get a specific menu item
         item = menu_db.get_item_by_id(item_id)
         if (not item):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
         
         return jsonify(item)
 
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def delete(self, item_id):
         # Delete a specific menu item
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
         if (not menu_db.delete_item(item_id)):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'status': 'success' })
 
@@ -95,24 +130,35 @@ class CreateMenuCategory(Resource):
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
-    @menu.expect(request_model.category_model)
+    @menu.response(500, 'Something went wrong')
+    @menu.expect(menu_model.category_request_model)
     def post(self):
+        # Create a new category in the menu
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
+        # Get and validate request body
         name = request.get_json().get('name')
         if (not name):
             abort(400, 'Missing required field \'name\'')
 
+        # Create new category in the database
         if (not menu_db.create_category(name)):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'status': 'success' })
 
     @jwt_required
-    @menu.response(200, 'Success', model=response_model.category_model)
-    @menu.response(400, 'Invalid Request')
+    @menu.response(200, 'Success', model=menu_model.category_response_model)
+    @menu.response(500, 'Something went wrong')
     def get(self):
+        # Get all the cateogies in the menu
         categories = menu_db.get_categories()
         if (not categories):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'categories': categories })
 
@@ -122,49 +168,55 @@ class MenuCategory(Resource):
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
-    @menu.expect(request_model.category_model)
+    @menu.response(500, 'Something went wrong')
+    @menu.expect(menu_model.category_request_model)
     def put(self, category_id):
+        # Edit a category
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
 
         edit = request.get_json()
-        editArr = []
-        editStatement = 'UPDATE category SET '
-
-        if (edit.get('name')):
-            editStatement += "name = %s, "
-            editArr.append(edit.get('name'))
-        if (edit.get('position')):
-            editStatement += "position = %s, "
-            editArr.append(edit.get('position'))
-
-        editStatement = editStatement.strip(', ') + ' WHERE id = %s'
-        editArr.append(category_id)
-        if (not menu_db.edit_category(editStatement, editArr)):
-            abort(400, 'Something went wrong')
+        # Edit the category in the database
+        if (not menu_db.edit_category(edit)):
+            abort(500, 'Something went wrong')
         
         return jsonify({ 'status': 'success' })
 
     @jwt_required
-    @menu.response(200, 'Success', model=response_model.specific_category_model)
-    @menu.response(400, 'Invalid Request')
+    @menu.response(200, 'Success', model=menu_model.category_items_response_model)
+    @menu.response(500, 'Something went wrong')
     def get(self, category_id):
+        # Get a category and its menu items
         category = menu_db.get_category(category_id)
         if (not category):
-            abort(400, 'Invalid request')
+            abort(500, 'Invalid request')
 
         return jsonify(category)
 
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def delete(self, category_id):
+        # Delete a category from the menu
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+    
         items = menu_db.get_items_by_category(category_id)
-        if (len(items) > 0):
+        if items is None:
+            abort(500, 'Something went wrong')
+        elif (len(items) > 0):
             abort(400, 'Can only delete empty category')
 
         # Delete a category
         if (not menu_db.delete_category(category_id)):
-            abort(400, 'Invalid request')
-
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'status': 'success' })
 
@@ -173,13 +225,21 @@ class ManuCategorySwap(Resource):
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def post(self, category_id1, category_id2):
+        # Swap position of two categories
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
         # Maintain the correct order when passing in arguments
         id1 = min(category_id1, category_id2)
         id2 = max(category_id1, category_id2)
 
         if (not menu_db.swapCategoryPositions(id1, id2)):
-            abort(400, 'Failed to swap category positions')
+            abort(500, 'Failed to swap category positions')
         
         return jsonify({ 'status': 'success' })
 
@@ -188,20 +248,34 @@ class MenuCategoryItem(Resource):
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def post(self, category_id, item_id):
         # Add an item to a category
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
         if (not menu_db.add_item_to_category(category_id, item_id)):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
         
         return jsonify({ 'status': 'success' })
 
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def delete(self, category_id, item_id):
         # Remove an item from a category
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
         if (not menu_db.remove_item_from_category(category_id, item_id)):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'status': 'success' })
 
@@ -211,25 +285,33 @@ class MenuCategoryItem(Resource):
 class CreateMenuIngredient(Resource):
     @jwt_required
     @menu.response(200, 'Success')
-    @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def get(self):
         # Get a list of all ingredients
         ingredients = menu_db.get_all_ingredients()
         if (not ingredients):
-            abort(400, 'Invalid request')
+            abort(500, 'Something went wrong')
         
         return ingredients
 
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
-    @menu.expect(request_model.ingredient_model)
+    @menu.expect(menu_model.ingredient_request_model)
     def post(self):
         # Create a new ingredient
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
+        # Validate request body
         name = request.get_json().get('name')
         if (not name):
             abort(400, 'Missing ingredient name')
 
+        # Create new ingredient item
         if (not menu_db.create_ingredient(name)):
             abort(400, 'Something went wrong')
 
@@ -238,40 +320,55 @@ class CreateMenuIngredient(Resource):
 @menu.route('/ingredient/<int:id>')
 class MenuIngredient(Resource):
     @jwt_required
-    @menu.response(200, 'Success', model=response_model.ingredient_model)
-    @menu.response(400, 'Invalid Request')
+    @menu.response(200, 'Success', model=menu_model.ingredient_response_model)
+    @menu.response(500, 'Something went wrong')
     def get(self, id):
         # Return a specific ingredient
         ingredient = menu_db.get_ingredient_by_id(id)
         if (not ingredient):
-            abort(400, 'Invalid request, ingredient not found')
+            abort(500, 'Something went wrong')
         
         return jsonify(ingredient)
 
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
-    @menu.expect(request_model.ingredient_model)
+    @menu.response(500, 'Something went wrong')
+    @menu.expect(menu_model.ingredient_request_model)
     def put(self, id):
         # Modify a specific ingredient
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
+        # Validate request body
         name = request.get_json().get('name')
         if (not name):
             abort(400, 'Invalid request')
 
+        # Edit ingredient in the menu
         if (not menu_db.edit_ingredient(name, id)):
-            abort(400, 'Something went wrong')
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'status': 'success' })
 
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def delete(self, id):
         # Delete a specific ingredient
         # Only if not used
 
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
         if (not menu_db.delete_ingredient(id)):
-            abort(400, 'Ingredient still in use')
+            abort(500, 'Ingredient still in use')
 
         return jsonify({ 'status': 'success' })
 
@@ -280,19 +377,35 @@ class MenuItemIngredient(Resource):
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def post(self, item_id, ingredient_id):
         # Add an ingredient to an item
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
+        # Add ingredient to item in the database
         if (not menu_db.add_ingredient_to_item(item_id, ingredient_id)):
-            abort(400, 'Something went wrong')
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'status': 'success' })
 
     @jwt_required
     @menu.response(200, 'Success')
     @menu.response(400, 'Invalid Request')
+    @menu.response(500, 'Something went wrong')
     def delete(self, item_id, ingredient_id):
         # Remove ingredient from an item
+
+        role = get_jwt_claims().get('role')
+        # Make sure user is a manager
+        if db.get_staff_title(role) != 'Manage':
+            abort(400, 'User is not a manager')
+
+        # Remove ingredeint from item in the database
         if (not menu_db.remove_ingredient_from_item(item_id, ingredient_id)):
-            abort(400, 'Something went wrong')
+            abort(500, 'Something went wrong')
 
         return jsonify({ 'status': 'success' })
